update task details
